+++ src/common/commands/generic_nack.rs
@@ -1,5 +1,4 @@
 use log::error;
-use num_traits::FromPrimitive;
 
 use crate::{CommandHeader, CommandId, SmppError, SmppReply};
 
@@ -64,8 +63,7 @@ impl generic_nack {
     }
 
     pub fn get_error(&self) -> SmppError {
-        FromPrimitive::from_u32(self.header.command_status)
-            .expect("Can not convert command_status to SmppError")
+        SmppError::from_command_status(self.header.command_status)
     }
 }