+++ src/common/commands/query_sm.rs
@@ -1,7 +1,6 @@
 use crate::common::parse_c_octet_string_nom;
 use crate::{CommandHeader, CommandId, SmppError, SmppReply};
 use nom::bytes::complete::take;
-use num_traits::FromPrimitive;
 
 #[derive(Debug, Clone)]
 pub struct query_sm {
@@ -142,8 +141,7 @@ impl query_sm_resp {
         self.header.command_status
     }
     pub fn get_error(&self) -> SmppError {
-        FromPrimitive::from_u32(self.header.command_status)
-            .expect("Can not convert command_status to SmppError")
+        SmppError::from_command_status(self.header.command_status)
     }
 
     pub fn decode(header: CommandHeader, pdu: &[u8]) -> Result<query_sm_resp, SmppError> {